+++ src/css/index.ts
@@ -1,16 +1,17 @@
 import createStyler from '../styler';
 import { Styler } from '../styler/types';
 import prefixer from './prefixer';
-import render from './render';
+import buildStyles, { aliasMap } from './render';
 import { isTransformProp } from './transform-props';
 import getValueType from './value-types';
 
 type Props = {
-  enableHardwareAcceleration?: boolean
+  enableHardwareAcceleration?: boolean,
+  preparseOutput?: boolean
 };
 
 const cssStyler = createStyler({
-  onRead: (key, { element }) => {
+  onRead: (key, { element, preparseOutput }) => {
     const valueType = getValueType(key);
 
     if (isTransformProp(key)) {
@@ -18,26 +19,21 @@ const cssStyler = createStyler({
         ? valueType.default || 0
         : 0;
     } else {
-      const domValue = window.getComputedStyle(element, null).getPropertyValue(prefixer(key)) || 0;
-      return (valueType && valueType.parse) ? valueType.parse(domValue) : domValue;
+      const domValue = window.getComputedStyle(element, null).getPropertyValue(prefixer(key, true)) || 0;
+      return (preparseOutput && valueType && valueType.parse) ? valueType.parse(domValue) : domValue;
     }
   },
   onRender: (state, { element, enableHardwareAcceleration }, changedValues) => {
-    render(element, state, changedValues, enableHardwareAcceleration);
+    element.style.cssText += buildStyles(state, changedValues, enableHardwareAcceleration);
   },
-  aliasMap: {
-    x: 'translateX',
-    y: 'translateY',
-    z: 'translateZ',
-    originX: 'transformOriginX',
-    originY: 'transformOriginY'
-  }
+  aliasMap
 });
 
 export default (element: HTMLElement, props?: Props): Styler => {
   return cssStyler({
     element,
     enableHardwareAcceleration: true,
+    preparseOutput: true,
     ...props
   });
 };