+++ src/css/render.ts
@@ -3,12 +3,22 @@ import prefixer from './prefixer';
 import getValueType from './value-types';
 import { State } from '../styler/types';
 
+export type AliasMap = { [key: string]: string };
+
+export const aliasMap: AliasMap = {
+  x: 'translateX',
+  y: 'translateY',
+  z: 'translateZ',
+  originX: 'transformOriginX',
+  originY: 'transformOriginY'
+};
+
 export default function buildStylePropertyString(
-  element: HTMLElement,
   state: State,
-  changedValues: string[],
-  enableHardwareAcceleration: boolean
+  changedValues: string[] | true = true,
+  enableHardwareAcceleration: boolean = true
  ) {
+  const valuesToChange = (changedValues === true) ? Object.keys(state) : changedValues;
   let propertyString = '';
   let transformString = '';
   let hasTransform = false;
@@ -16,9 +26,9 @@ export default function buildStylePropertyString(
 
   // First check if there are any changed transform values
   // and if true add all transform values
-  const numChangedValues = changedValues.length;
+  const numChangedValues = valuesToChange.length;
   for (let i = 0; i < numChangedValues; i++) {
-    const key = changedValues[i];
+    const key = valuesToChange[i];
 
     // If this is a transform property, add all other transform props
     // to changedValues and then break
@@ -26,8 +36,8 @@ export default function buildStylePropertyString(
       hasTransform = true;
 
       for (const stateKey in state) {
-        if (isTransformProp(stateKey) && changedValues.indexOf(stateKey) === -1) {
-          changedValues.push(stateKey);
+        if (isTransformProp(stateKey) && valuesToChange.indexOf(stateKey) === -1) {
+          valuesToChange.push(stateKey);
         }
       }
 
@@ -35,13 +45,13 @@ export default function buildStylePropertyString(
     }
   }
 
-  changedValues.sort(sortTransformProps);
+  valuesToChange.sort(sortTransformProps);
 
   // Now run through each property, and decide which is a plain style props,
   // a transform property and CSS vars and handle accordingly
-  const totalNumChangedValues = changedValues.length;
+  const totalNumChangedValues = valuesToChange.length;
   for (let i = 0; i < totalNumChangedValues; i++) {
-    const key = changedValues[i];
+    const key = valuesToChange[i];
     let value: any = state[key];
 
     // If this is a number or object and we have filter, apply filter
@@ -70,5 +80,5 @@ export default function buildStylePropertyString(
     propertyString += ';' + prefixer('transform', true) + ':' + transformString;
   }
 
-  element.style.cssText += propertyString;
+  return propertyString;
 }